Fully functional CLI mode added.
usage:
	java -jar JWiiload.jar <address> <filename> <application arguments>
	
	pass $WIILOAD as the first argument if your environment is set up that way.
		 "AUTO" as the first argument to try to automatically find the Wii.
		 "PREV" as the first argument to use the last known Wii IP that worked.

Field for Wii's IP and send button also added to GUI, but not yet coded.

+++ JWiiLoad/src/JWiiLoad.java
@@ -5,6 +5,8 @@ import java.util.zip.*;
 import java.awt.Container;
 import java.awt.Dimension;
 import java.awt.FlowLayout;
+//import java.awt.GridLayout;
+import java.awt.TextField;
 import java.io.BufferedInputStream;
 import java.io.DataOutputStream;
 import java.io.FileInputStream;
@@ -53,21 +55,103 @@ public class JWiiLoad {
 	static String host;
 	static String ip;
 
+	static TextField wiiip = new TextField("Enter Wii IP");
+
 	static int startip = prefs.getInt("start", 0);
-	static String lastip = prefs.get("host", "Enter IP here");
+	static String lastip = prefs.get("host", "null");
 	static boolean autosend = prefs.getBoolean("auto", true);
 	static int timeout = prefs.getInt("timeout",100);
+
+	static String[] gargs;
 
 	public static void main(String[] args) 
 	{
 		//button5.setEnabled(false);
+		gargs = args;
 
-		do{
-			fileselect.showOpenDialog(null);
-			filename = fileselect.getSelectedFile();
-		}while(filename==null);
+		if (args.length==1)
+		{
+			System.out.println("JWiiload .9\ncoded by VGMoose, based on wiiload by dhewg\n\nusage:\n\tjava -jar JWiiload.jar <address> <filename> <application arguments>\n\npass $WIILOAD as the first argument if your environment is set up that way.\n\npass \"AUTO\" as the first argument to try to automatically find the Wii.\n\npass \"PREV\" as the first argument to use the last known Wii IP that worked.\n");
+			System.exit(27);
+		}
+		if (args.length!=0)
+		{
+			System.out.println("Welcome to JWiiload CLI!");
+			
+			filename = new File(args[1]);
+			if (!filename.exists())
+			{
+				System.out.println("File at "+filename.getAbsolutePath()+" not found!");
+				System.exit(2);
+			}
 
-		System.out.println("".length());
+			if (args[0].startsWith("tcp:"))
+				args[0]=args[0].substring(4);
+
+			if (args[0].equals("PREV"))
+			{
+				if (lastip.equals("null"))
+				{
+					System.out.println("There is no known previous working IP stored in this machine!");
+					args[0]="AUTO";
+				}
+				else
+					args[0]=lastip;
+			}
+
+			if (args[0].equals("AUTO"))
+			{	
+				tripleScan();
+				if (host==null)
+				{
+					System.out.println("Could not find Wii through auto-detection");
+					System.exit(3);
+				}
+				else if (host.equals("rate"))
+				{
+					System.out.println("Too many wireless requests to auto-detect, try again later.");
+					System.exit(4);
+				}
+				else
+					args[0]=host;
+			}
+
+
+			System.out.println("\nIP: "+args[0]);
+			host = args[0];
+
+			System.out.println("File: "+filename.getName());
+
+			if (args.length>2)
+				for (int x=2;x<args.length;x++)
+				{
+					arguments+=args[x];
+					if (x!=args.length-1)
+						arguments+=" ";
+				}
+
+			System.out.print("Arguments: "+arguments);
+			
+			if (arguments.length()==0)
+				System.out.print("none");
+			
+			System.out.println("\n");
+			
+			if (socket==null)
+				if (!connects())
+					System.exit(1);
+
+		}
+		else
+		{
+			do{
+				fileselect.showOpenDialog(null);
+				filename = fileselect.getSelectedFile();
+			}while(filename==null);
+
+			createWindow();		// Create the JFrame GUI
+
+		}
 
 		if (filename!=null)
 		{
@@ -75,34 +159,54 @@ public class JWiiLoad {
 			button5.setText("Send "+filename.getName());
 		}
 
-		createWindow();		// Create the JFrame GUI
-
 		if (filename!=null)
 			compressData();	
 
 		Runtime.getRuntime().addShutdownHook(new Thread(new Runnable() {
 			public void run() {
-				compressed.delete();
+				if (compressed!=filename)
+					compressed.delete();
 			}
 		}));
 
-		tripleScan();
+		if (args.length==0)
+			tripleScan();
 
 		if (filename!=null)
 			wiisend();
 
 	}
 
+	public static boolean connects()
+	{
+		System.out.println("Connecting to "+host+"...");
+		
+		try{
+			socket = new Socket(host, port);
+			System.out.println("Connection successful!\n");
+			return true;
+		} catch (Exception e) {
+			System.out.println("Connection failed.");
+			return false;
+		}
+		
+	}
+	
 	public static void compressData()
 	{
 		try
 		{
 			// Compress the file to send it faster
 			text1.setText("Compressing data...");
+			System.out.println("Compressing data...");
 			compressed = compressFile(filename);
 			text1.setText("Data compressed!");	//+ (int)(100*((compressed.length()+0.0)/filename.length()))+"% smaller)");
+			System.out.println("Compression successful! ("+(int)(100*((compressed.length()+0.0)/filename.length()))+"% smaller)\n");
+
 		} catch(Exception e){
-			e.printStackTrace();
+			// Fall back in case compressed file can't be written
+			System.out.println("Compression failed! Not going to send it compressed.\n");
+			compressed = filename;
 		}
 	}
 
@@ -122,6 +226,8 @@ public class JWiiLoad {
 		try
 		{
 			// Open socket to wii with host and port and setup output stream
+			System.out.println("Greeting the Wii...");
+
 			if (host==null)
 				socket = new Socket(host, port);
 
@@ -131,6 +237,7 @@ public class JWiiLoad {
 			DataOutputStream dos = new DataOutputStream(os);
 
 			text1.setText("Preparing data...");
+			System.out.println("Preparing local data...");
 
 			byte max = 0;
 			byte min = 5;
@@ -148,6 +255,7 @@ public class JWiiLoad {
 			int numRead=0;
 
 			text1.setText("Talking to Wii...");
+			System.out.println("Perparing remote data...");
 
 			dos.writeBytes("HAXX");
 
@@ -162,17 +270,20 @@ public class JWiiLoad {
 			//dos.size();	// Number of bytes sent so far, should be 16
 
 			text1.setText("Sending "+filename.getName()+"...");
+			System.out.println("Sending "+filename.getName()+"...");
 			dos.flush();
 
 			while ( ( numRead=bis.read(b)) > 0) {
-
 				dos.write(b,0,numRead);
 				dos.flush();
-
 			}
 			dos.flush();
 
 			text1.setText("Talking to Wii...");
+			if (arguments.length()!=0)
+				System.out.println("Sending arguments...");
+			else
+				System.out.println("Finishing up...");
 
 			dos.writeBytes(filename.getName()+"\0");
 
@@ -182,8 +293,10 @@ public class JWiiLoad {
 				dos.writeBytes(x+"\0");
 
 			text1.setText("All done!");
+			System.out.println("\nFile transfer successful!");
 
-			compressed.delete();
+			if (compressed!=filename)
+				compressed.delete();
 
 
 		}
@@ -193,11 +306,22 @@ public class JWiiLoad {
 			String[] selections = {"Retry","Stop"};
 			int a=0;
 
-			if (host.equals("rate"))
-				a = JOptionPane.showOptionDialog(frame,"Rate Limit Exceeded.\nPlease wait a little while, then try again.","Error", JOptionPane.ERROR_MESSAGE, 0, null, selections, null);
-			else
-				a = JOptionPane.showOptionDialog(frame,"No Wii found.\nIs the Homebrew Channel running?","Error",JOptionPane.ERROR_MESSAGE, 0, null,selections,null);
+			if (host==null)
+				host="";
 
+			if (gargs.length==0)
+			{
+				if (host.equals("rate"))
+					a = JOptionPane.showOptionDialog(frame,"Rate Limit Exceeded.\nPlease wait a little while, then try again.","Error", JOptionPane.ERROR_MESSAGE, 0, null, selections, null);
+				else
+					a = JOptionPane.showOptionDialog(frame,"No Wii found.\nIs the Homebrew Channel running?","Error",JOptionPane.ERROR_MESSAGE, 0, null,selections,null);
+			}
+			else
+			{
+				System.out.println("No Wii found at "+host+"!");
+				System.exit(1);
+			}
+			
 			if (a==0)
 			{
 				tripleScan();
@@ -205,17 +329,6 @@ public class JWiiLoad {
 			}
 
 		}
-//		catch (Exception ex) {
-//			text1.setText("No Wii found");
-//			ex.printStackTrace();
-//			if (host.equals("rate"))
-//			{
-//				text1.setText("Rate Limited");
-//				JOptionPane.showMessageDialog(frame,"Rate Limit Exceeded.\nPlease wait a little while, then try again.");
-//				tripleScan();
-//				wiisend();
-//			}
-//		}
 	}
 
 	static void scan(int t)
@@ -223,6 +336,7 @@ public class JWiiLoad {
 		host=null;
 
 		text1.setText("Finding Wii...");
+		System.out.println("Searching for a Wii...");
 		String output = null;
 
 		InetAddress localhost=null;
@@ -246,17 +360,18 @@ public class JWiiLoad {
 				if (address.isReachable(10*t))
 				{
 					output = address.toString().substring(1);
-					System.out.println(output + " is on the network");
+					System.out.print(output + " is on the network");
 
 					// Attempt to open a socket
 					try
 					{
 						socket = new Socket(output,port);
-						System.out.println("And is potentially a Wii!");
+						System.out.println("and is potentially a Wii!");
 						text1.setText("Wii found!");
 						host=output;
 						return;
 					} catch (Exception e) {
+						System.out.println();
 						//e.printStackTrace();
 					}
 
@@ -281,17 +396,25 @@ public class JWiiLoad {
 	private static void createWindow() {
 
 		frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE); 
-		//text1.setPreferredSize(new Dimension(200, 20));
-		text1.setPreferredSize(new Dimension(200, 200));
+		text1.setPreferredSize(new Dimension(200, 20));
 		//button1.setSize(100,100);
 		Container content = frame.getContentPane();
+
+
 		FlowLayout fl = new FlowLayout();
-		content.setLayout(fl); 
+
+		content.setLayout(fl);
+
+		content.add(text1);
+
+		content.add(wiiip);
+		content.add(button5);
 
 		//	content.add(button1);
 		//	content.add(button2);
 
-		content.add(text1);
+
+
 
 		//	content.add(button5);
